# perf: Reduce template compilation time by ~16%