--- a/constraint/pkg/regorewriter/regorewriter.go
+++ b/constraint/pkg/regorewriter/regorewriter.go
@@ -61,12 +61,7 @@ func New(pt PackageTransformer, libs []string, externs []string) (*RegoRewriter,
 }
 
 // add is the internal method for parsing a module and entering it into the bookkeeping.
-func (r *RegoRewriter) add(path, src string, slice *[]*Module) error {
-	m, err := ast.ParseModule(path, src)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrInvalidModule, err)
-	}
-
+func (r *RegoRewriter) add(path string, m *ast.Module, slice *[]*Module) error {
 	r.addModule(path, m, slice)
 	return nil
 }
@@ -81,13 +76,13 @@ func (r *RegoRewriter) AddEntryPointModule(path string, m *ast.Module) {
 
 // AddEntryPoint adds a base source which will not have it's package path rewritten.  These correspond
 // to the rego that will be populated into a ConstraintTemplate with the 'violation' rule.
-func (r *RegoRewriter) AddEntryPoint(path, src string) error {
-	return r.add(path, src, &r.entryPoints)
+func (r *RegoRewriter) AddEntryPoint(path string, m *ast.Module) error {
+	return r.add(path, m, &r.entryPoints)
 }
 
 // AddLib adds a library source which will have the package path updated.
-func (r *RegoRewriter) AddLib(path, src string) error {
-	return r.add(path, src, &r.libs)
+func (r *RegoRewriter) AddLib(path string, m *ast.Module) error {
+	return r.add(path, m, &r.libs)
 }
 
 // addTestDir adds a test dir inside one of the provided paths.
@@ -128,7 +123,13 @@ func (r *RegoRewriter) addFileFromFs(path string, slice *[]*Module) error {
 	if err != nil {
 		return fmt.Errorf("%w: %v", ErrReadingFile, err)
 	}
-	return r.add(path, string(bytes), slice)
+
+	m, err := ast.ParseModule(path, string(bytes))
+	if err != nil {
+		return err
+	}
+
+	return r.add(path, m, slice)
 }
 
 // addPathFromFs adds a module from the local filesystem.