@willbeason 2022-03-29
Ultimately this is just refactoring - no visible changes to users or consumers of this library.