@willbeason 2022-04-01
Tested locally via go.work - passes in Gatekeeper.